pyproject.toml
# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db --import-mode=importlib"
python_files = [
    "tests.py",
    "test_*.py",
]

# ==== Coverage ====
[tool.coverage.run]
include = ["magazix/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]

# ==== mypy ====
[tool.mypy]
python_version = "3.13"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/djlint/djLint/issues/687
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

[tool.ruff]
# Exclude a variety of commonly ignored directories.
extend-exclude = [
    "*/migrations/*.py",
    "staticfiles/*",
]

[tool.ruff.lint]
select = [
    "F",
    "E",
    "W",
    "C90",
    "I",
    "N",
    "UP",
    "YTT",
    # "ANN", # flake8-annotations: we should support this in the future but 100+ errors atm
    "ASYNC",
    "S",
    "BLE",
    "FBT",
    "B",
    "A",
    "COM",
    "C4",
    "DTZ",
    "T10",
    "DJ",
    "EM",
    "EXE",
    "FA",
    'ISC',
    "ICN",
    "G",
    'INP',
    'PIE',
    "T20",
    'PYI',
    'PT',
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SLOT",
    "SIM",
    "TID",
    "TC",
    "INT",
    # "ARG", # Unused function argument
    "PTH",
    "ERA",
    "PD",
    "PGH",
    "PL",
    "TRY",
    "FLY",
    # "NPY",
    # "AIR",
    "PERF",
    # "FURB",
    # "LOG",
    "RUF",
]
ignore = [
    "S101", # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
    "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
    "SIM102", # sometimes it's better to nest
    # of types for comparison.
    # Deactivated because it can make the code slow:
    # https://github.com/astral-sh/ruff/issues/7871
]

[tool.ruff.lint.isort]
force-single-line = true

[dependency-groups]
dev = [
    "coverage==7.13.1",
    "django-coverage-plugin==3.2.0",
    "django-debug-toolbar==6.2.0",
    "django-extensions==4.1",
    "django-stubs[compatible-mypy]==5.2.9",
    "djlint==1.36.4",
    "factory-boy==3.3.2",
    "ipdb==0.13.13",
    "mypy==1.19.1",
    "pre-commit==4.5.1",
    "psycopg[c]==3.3.2",
    "pytest==9.0.2",
    "pytest-django==4.11.1",
    "pytest-sugar==1.1.1",
    "ruff==0.14.14",
    "sphinx==9.1.0",
    "sphinx-autobuild==2025.8.25",
    "watchfiles==1.1.1",
    "werkzeug[watchdog]==3.1.5",
]

[project]
name = "magazix"
version = "0.1.0"
description = "Magazix is a lightweight, server-rendered e-commerce platform built with Django. It uses Django Templates for SSR, HTMX for partial page updates, and Alpine.js for small UI interactions - avoiding SPA complexity while keeping the interface responsive and fast. The project follows a modular, app-based architecture inspired by classic platforms like OpenCart, focusing on simplicity, SEO-friendliness, and maintainability."
readme = "README.md"

license = { text = "MIT" }
authors = [
  { name = "Maxim Kozlov", email = "maxim-kozlov@magazix.app" },
]
requires-python = "==3.13.*"
dependencies = [
    "argon2-cffi==25.1.0",
    "celery==5.6.2",
    "crispy-tailwind>=1.0.3",
    "django==5.2.10",
    "django-allauth[mfa]==65.14.0",
    "django-anymail==14.0",
    "django-celery-beat==2.8.1",
    "django-crispy-forms==2.5",
    "django-environ==0.12.0",
    "django-meta>=2.5.0",
    "django-mptt>=0.18.0",
    "django-model-utils==5.0.0",
    "django-mptt>=0.18.0",
    "django-redis==6.0.0",
    "django-template-partials>=25.3",
    "flower==2.0.1",
    "gunicorn==24.1.1",
    "hiredis==3.3.0",
    "pillow==12.1.0",
    "psycopg[c]==3.3.2",
    "python-slugify==8.0.4",
    "redis==7.1.0",
    "sentry-sdk==2.50.0",
    "whitenoise==6.11.0",
]
